@@ -36,8 +36,8 @@ const ambientZone = Zone.current;
 // inside of a `describe` but outside of a `beforeEach` or `it`.
 const syncZone = ambientZone.fork(new SyncTestZoneSpec('jest.describe'));
 function wrapDescribeInZone(describeBody) {
-  return function() {
-    return syncZone.run(describeBody, null, arguments);
+  return function(...args) {
+    return syncZone.run(describeBody, null, args);
   };
 }
 
@@ -48,61 +48,82 @@ function wrapTestInZone(testBody) {
   if (testBody === undefined) {
     return;
   }
-  return testBody.length === 0
-    ? () => testProxyZone.run(testBody, null)
-    : done => testProxyZone.run(testBody, null, [done]);
+
+  const wrappedFunc = function() {
+    return testProxyZone.run(testBody, null, arguments);
+  };
+  try {
+    Object.defineProperty(wrappedFunc, 'length', {
+      configurable: true,
+      writable: true,
+      enumerable: false
+    });
+    wrappedFunc.length = testBody.length;
+  } catch (e) {
+    return testBody.length === 0
+      ? () => testProxyZone.run(testBody, null)
+      : done => testProxyZone.run(testBody, null, [done]);
+  }
+  return wrappedFunc;
 }
 
+/**
+ * bind describe method to wrap describe.each function
+ */
 const bindDescribe = originalJestFn =>
-  function() {
-    const eachArguments = arguments;
-    return function(description, specDefinitions, timeout) {
-      arguments[1] = wrapDescribeInZone(specDefinitions);
-      return originalJestFn.apply(this, eachArguments).apply(this, arguments);
+  function(...eachArgs) {
+    return function(...args) {
+      args[1] = wrapDescribeInZone(args[1]);
+      return originalJestFn.apply(this, eachArgs).apply(this, args);
+    };
+  };
+
+/**
+ * bind test method to wrap test.each function
+ */
+const bindTest = originalJestFn =>
+  function(...eachArgs) {
+    return function(...args) {
+      args[1] = wrapTestInZone(args[1]);
+      return originalJestFn.apply(this, eachArgs).apply(this, args);
     };
   };
 
 ['xdescribe', 'fdescribe', 'describe'].forEach(methodName => {
   const originaljestFn = env[methodName];
-  env[methodName] = function(description, specDefinitions, timeout) {
-    arguments[1] = wrapDescribeInZone(specDefinitions);
-    return originaljestFn.apply(this, arguments);
+  env[methodName] = function(...args) {
+    args[1] = wrapDescribeInZone(args[1]);
+    return originaljestFn.apply(this, args);
   };
   env[methodName].each = bindDescribe(originaljestFn.each);
   if (methodName === 'describe') {
     env[methodName].only = env['fdescribe'];
     env[methodName].skip = env['xdescribe'];
-    env[methodName].only.each = bindDescribe(originaljestFn.only.each);
-    env[methodName].skip.each = bindDescribe(originaljestFn.skip.each);
   }
 });
 
 ['xit', 'fit', 'xtest', 'test', 'it'].forEach(methodName => {
   const originaljestFn = env[methodName];
-  env[methodName] = function(description, specDefinitions, timeout) {
-    arguments[1] = wrapTestInZone(specDefinitions);
-    return originaljestFn.apply(this, arguments);
+  env[methodName] = function(...args) {
+    args[1] = wrapTestInZone(args[1]);
+    return originaljestFn.apply(this, args);
   };
-  // The revised method will be populated to the final each method, so we only declare the method that in the new globals
-  env[methodName].each = originaljestFn.each;
+  env[methodName].each = bindTest(originaljestFn.each);
 
   if (methodName === 'test' || methodName === 'it') {
     env[methodName].only = env['fit'];
-    env[methodName].only.each = originaljestFn.only.each;
-
     env[methodName].skip = env['xit'];
-    env[methodName].skip.each = originaljestFn.skip.each;
 
-    env[methodName].todo = function(description) {
-      return originaljestFn.todo.apply(this, arguments);
+    env[methodName].todo = function(...args) {
+      return originaljestFn.todo.apply(this, args);
     };
   }
 });
 
 ['beforeEach', 'afterEach', 'beforeAll', 'afterAll'].forEach(methodName => {
   const originaljestFn = env[methodName];
-  env[methodName] = function(specDefinitions, timeout) {
-    arguments[0] = wrapTestInZone(specDefinitions);
-    return originaljestFn.apply(this, arguments);
+  env[methodName] = function(...args) {
+    args[0] = wrapTestInZone(args[0]);
+    return originaljestFn.apply(this, args);
   };
 });
